requirements.txt
tinkoff-investments==0.2.0b27
numpy==1.22.4
